requirements.txt
anyio==4.10.0
assemblyai==0.40.2
astor==0.8.1
Bio==1.8.0
chunkr_ai==0.3.7
colorama==0.4.6
colorlog==6.9.0
crawl4ai==0.7.4
datasets==4.0.0
docker_py==1.10.6
docx2markdown==0.1.1
e2b_code_interpreter==2.0.0
ffmpeg==1.4
google_search_results==2.4.2
httpx==0.28.1
jupyter_client==8.6.3
loguru==0.7.3
mcp==1.13.1
nest_asyncio==1.6.0
numpy==2.3.3
openai==1.107.1
opencv_python==4.11.0.86
openpyxl==3.1.5
pandas==2.3.2
pdfminer==20191125
Pillow==11.3.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
python_pptx==1.0.2
Requests==2.32.5
retry==0.9.2
rich==14.1.0
scenedetect==0.6.6
serpapi==0.1.5
tenacity==9.1.2
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
xls2xlsx==0.2.0
xmltodict==0.14.2
yt_dlp==2025.9.5
tabulate==0.9.0
